Project for harvesting metadata from DigitalArkivet
The Digital Archives of Norway also known as DigitalArkivet. The Digital
Archives is mainly an internet site, a publishing channel for the
National Archives of Norway. You'll find sources/documents. Every
digital and digitalized has metadata which DIS-Norge, The Genealogy
Society of Norway-DIS, wants to reuse. Our slogan is "we make genealogy
easier"